Move education group details in new file
(cherry picked from commit 07d1251)
