+++ base/urls.py
@@ -29,6 +29,7 @@ from django.conf.urls import url, include
 from django.conf.urls.static import static
 
 import base.views.education_groups.create
+import base.views.education_groups.detail
 import base.views.education_groups.search
 import base.views.learning_units.common
 import base.views.learning_units.create
@@ -227,10 +228,11 @@ urlpatterns = [
         url(r'^new/(?P<parent_id>[0-9]+)/$', base.views.education_groups.create.create_education_group,
             name='new_education_group'),
         url(r'^(?P<education_group_year_id>[0-9]+)/', include([
-            url(r'^$', education_group.EducationGroupRead.as_view(), name='education_group_read'),
+            url(r'^$', base.views.education_groups.detail.EducationGroupRead.as_view(), name='education_group_read'),
             url(r'^update/$', update_education_group, name="update_education_group"),
-            url(r'^diplomas/$', education_group.EducationGroupDiplomas.as_view(), name='education_group_diplomas'),
-            url(r'^informations/$', education_group.education_group_general_informations,
+            url(r'^diplomas/$', base.views.education_groups.detail.EducationGroupDiplomas.as_view(),
+                name='education_group_diplomas'),
+            url(r'^informations/$', base.views.education_groups.detail.EducationGroupGeneralInformation.as_view(),
                 name='education_group_general_informations'),
             url(r'^informations/edit/$', education_group.education_group_year_pedagogy_edit,
                 name="education_group_pedagogy_edit"),
@@ -244,10 +246,12 @@ urlpatterns = [
                 name="education_group_pedagogy_get_terms"),
 
             url(r'^administrative/', include([
-                url(u'^$', education_group.education_group_administrative_data, name='education_group_administrative'),
+                url(u'^$', base.views.education_groups.detail.EducationGroupAdministrativeData.as_view(),
+                    name='education_group_administrative'),
                 url(u'^edit/$', education_group.education_group_edit_administrative_data,
                     name='education_group_edit_administrative')])),
-            url(r'^content/$', education_group.education_group_content, name='education_group_content'),
+            url(r'^content/$', base.views.education_groups.detail.EducationGroupContent.as_view(),
+                name='education_group_content'),
             url(r'^admission_conditions/$',
                 education_group.education_group_year_admission_condition_edit,
                 name='education_group_year_admission_condition_edit'),
