+++ base/views/education_group.py
@@ -24,185 +24,38 @@
 #
 ##############################################################################
 import json
-from collections import OrderedDict
 
 from ckeditor.widgets import CKEditorWidget
 from django import forms
-from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required, permission_required
-from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.core.exceptions import PermissionDenied
 from django.db.models import Prefetch
 from django.http import HttpResponseRedirect, JsonResponse
 from django.shortcuts import get_object_or_404, redirect
 from django.urls import reverse
-from django.utils.decorators import method_decorator
 from django.utils.translation import ugettext_lazy as _
 from django.views.decorators.http import require_http_methods
-from django.views.generic import DetailView
 from waffle.decorators import waffle_flag
 
 from base import models as mdl
 from base.business import education_group as education_group_business
 from base.business.education_group import assert_category_of_education_group_year
 from base.business.learning_unit import find_language_in_settings
-from base.forms.education_group_general_informations import EducationGroupGeneralInformationsForm
 from base.forms.education_group_pedagogy_edit import EducationGroupPedagogyEditForm
 from base.forms.education_groups_administrative_data import CourseEnrollmentForm, AdministrativeDataFormset
 from base.models.admission_condition import AdmissionConditionLine, AdmissionCondition
 from base.models.education_group_year import EducationGroupYear
 from base.models.enums import academic_calendar_type
 from base.models.enums import education_group_categories
-from base.models.person import Person
+from base.views.education_groups.detail import get_education_group_root
 from base.views.learning_units.common import get_text_label_translated
-from cms import models as mdl_cms
 from cms.enums import entity_name
 from cms.models.text_label import TextLabel
 from cms.models.translated_text import TranslatedText
 from cms.models.translated_text_label import TranslatedTextLabel
 from osis_common.decorators.ajax import ajax_required
 from . import layout
-
-CODE_SCS = 'code_scs'
-TITLE = 'title'
-CREDITS_MIN = "credits_min"
-CREDITS_MAX = "credits_max"
-BLOCK = "block"
-SESSIONS_DEROGATION = "sessions_derogation"
-NUMBER_SESSIONS = 3
-
-
-@method_decorator(login_required, name='dispatch')
-class EducationGroupGenericDetailView(PermissionRequiredMixin, DetailView):
-    # DetailView
-    model = EducationGroupYear
-    context_object_name = "education_group_year"
-    pk_url_kwarg = 'education_group_year_id'
-
-    # PermissionRequiredMixin
-    permission_required = 'base.can_access_education_group'
-    raise_exception = True
-
-    limited_by_category = None
-
-    def get_person(self):
-        return get_object_or_404(Person, user=self.request.user)
-
-    def get_root(self):
-        return self.request.GET.get("root")
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        # This objects are mandatory for all education group views
-        context['person'] = self.get_person()
-        context['root'] = self.get_root()
-        context['parent'] = _get_education_group_root(self.get_root(), self.object)
-        return context
-
-    def get(self, request, *args, **kwargs):
-        if self.limited_by_category:
-            assert_category_of_education_group_year(self.get_object(), self.limited_by_category)
-        return super().get(request, *args, **kwargs)
-
-
-class EducationGroupRead(EducationGroupGenericDetailView):
-    templates = {
-        education_group_categories.TRAINING: "education_group/identification_training_details.html",
-        education_group_categories.MINI_TRAINING: "education_group/identification_mini_training_details.html",
-        education_group_categories.GROUP: "education_group/identification_group_details.html"
-    }
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        # TODO Use value_list
-        context["education_group_languages"] = [
-            education_group_language.language.name for education_group_language in
-            mdl.education_group_language.find_by_education_group_year(self.object)
-        ]
-        context['enums'] = mdl.enums.education_group_categories
-
-        return context
-
-    def get_template_names(self):
-        return self.templates.get(self.object.education_group_type.category)
-
-
-class EducationGroupDiplomas(EducationGroupGenericDetailView):
-    template_name = "education_group/tab_diplomas.html"
-    limited_by_category = (education_group_categories.TRAINING,)
-
-
-@login_required
-@permission_required('base.can_access_education_group', raise_exception=True)
-def education_group_general_informations(request, education_group_year_id):
-    education_group_year = get_object_or_404(EducationGroupYear, id=education_group_year_id)
-    assert_category_of_education_group_year(
-        education_group_year, (education_group_categories.TRAINING, education_group_categories.MINI_TRAINING))
-
-    cms_label = mdl_cms.translated_text.find_labels_list_by_label_entity_and_reference(entity_name.OFFER_YEAR,
-                                                                                       education_group_year_id)
-
-    fr_language = next((lang for lang in settings.LANGUAGES if lang[0] == 'fr-be'), None)
-    en_language = next((lang for lang in settings.LANGUAGES if lang[0] == 'en'), None)
-
-    education_group_year_root_id = request.GET.get('root')
-    parent = _get_education_group_root(education_group_year_root_id, education_group_year)
-
-    context = {
-        'parent': parent,
-        'can_edit_information': request.user.has_perm('base.can_edit_educationgroup_pedagogy'),
-        'education_group_year': education_group_year,
-        'cms_labels_translated': _get_cms_label_data(cms_label,
-                                                     mdl.person.get_user_interface_language(request.user)),
-        'form_french': EducationGroupGeneralInformationsForm(education_group_year=education_group_year,
-                                                             language=fr_language, text_labels_name=cms_label),
-        'form_english': EducationGroupGeneralInformationsForm(education_group_year=education_group_year,
-                                                              language=en_language, text_labels_name=cms_label)
-    }
-    return layout.render(request, "education_group/tab_general_informations.html", context)
-
-
-def _get_cms_label_data(cms_label, user_language):
-    cms_label_data = OrderedDict()
-    translated_labels = mdl_cms.translated_text_label.search(text_entity=entity_name.OFFER_YEAR,
-                                                             labels=cms_label,
-                                                             language=user_language)
-    for label in cms_label:
-        translated_text = next((trans.label for trans in translated_labels if trans.text_label.label == label), None)
-        cms_label_data[label] = translated_text
-    return cms_label_data
-
-
-@login_required
-@permission_required('base.can_access_education_group', raise_exception=True)
-def education_group_administrative_data(request, education_group_year_id):
-    education_group_year = get_object_or_404(EducationGroupYear, pk=education_group_year_id)
-
-    assert_category_of_education_group_year(education_group_year, (education_group_categories.TRAINING,))
-
-    education_group_year_root_id = request.GET.get('root')
-    parent = _get_education_group_root(education_group_year_root_id, education_group_year)
-
-    context = {'parent': parent,
-               'education_group_year': education_group_year,
-               'course_enrollment': get_dates(academic_calendar_type.COURSE_ENROLLMENT, education_group_year),
-               'mandataries': mdl.mandatary.find_by_education_group_year(education_group_year),
-               'pgm_mgrs': mdl.program_manager.find_by_education_group(education_group_year.education_group)}
-    context.update({'exam_enrollments': get_sessions_dates(academic_calendar_type.EXAM_ENROLLMENTS,
-                                                           education_group_year)})
-    context.update({'scores_exam_submission': get_sessions_dates(academic_calendar_type.SCORES_EXAM_SUBMISSION,
-                                                                 education_group_year)})
-    context.update({'dissertation_submission': get_sessions_dates(academic_calendar_type.DISSERTATION_SUBMISSION,
-                                                                  education_group_year)})
-    context.update({'deliberation': get_sessions_dates(academic_calendar_type.DELIBERATION,
-                                                       education_group_year)})
-    context.update({'scores_exam_diffusion': get_sessions_dates(academic_calendar_type.SCORES_EXAM_DIFFUSION,
-                                                                education_group_year)})
-    context.update({"can_edit_administrative_data": education_group_business.can_user_edit_administrative_data(
-        request.user, education_group_year)})
-    return layout.render(request, "education_group/tab_administrative_data.html", context)
 
 
 @login_required
@@ -236,101 +89,6 @@ def education_group_edit_administrative_data(request, education_group_year_id):
         return HttpResponseRedirect(reverse('education_group_administrative', args=(education_group_year_id,)))
 
     return layout.render(request, "education_group/tab_edit_administrative_data.html", locals())
-
-
-def get_sessions_dates(an_academic_calendar_type, an_education_group_year):
-    date_dict = {}
-
-    for session_number in range(NUMBER_SESSIONS):
-        session = mdl.session_exam_calendar.get_by_session_reference_and_academic_year(
-            session_number + 1,
-            an_academic_calendar_type,
-            an_education_group_year.academic_year)
-        if session:
-            dates = mdl.offer_year_calendar.get_by_education_group_year_and_academic_calendar(session.academic_calendar,
-                                                                                              an_education_group_year)
-            date_dict['session{}'.format(session_number+1)] = dates
-
-    return date_dict
-
-
-def get_dates(an_academic_calendar_type, an_education_group_year):
-    ac = mdl.academic_calendar.get_by_reference_and_academic_year(an_academic_calendar_type,
-                                                                  an_education_group_year.academic_year)
-    if ac:
-        dates = mdl.offer_year_calendar.get_by_education_group_year_and_academic_calendar(ac, an_education_group_year)
-        return {'dates': dates}
-    else:
-        return {}
-
-
-@login_required
-@permission_required('base.can_access_education_group', raise_exception=True)
-def education_group_content(request, education_group_year_id):
-    return _education_group_content_tab(request, education_group_year_id)
-
-
-def _education_group_content_tab(request, education_group_year_id):
-    education_group_year = mdl.education_group_year.find_by_id(education_group_year_id)
-    education_group_year_root_id = request.GET.get('root')
-    parent = _get_education_group_root(education_group_year_root_id, education_group_year)
-
-    context = {'parent': parent,
-               'education_group_year': education_group_year,
-               'group_elements': _group_elements(education_group_year),
-               }
-    return layout.render(request, "education_group/tab_content.html", context)
-
-
-def _get_education_group_root(education_group_year_root_id, default_education_group_year_root):
-    return get_object_or_404(mdl.education_group_year.EducationGroupYear, id=education_group_year_root_id) \
-        if education_group_year_root_id else default_education_group_year_root
-
-
-def _group_elements(education_group_yr):
-    group_elements = mdl.group_element_year.find_by_parent(education_group_yr)
-    if group_elements:
-        return _get_group_elements_data(group_elements)
-
-    return None
-
-
-def _get_group_elements_data(group_elements):
-    group_elements_data = []
-    for group_element in group_elements:
-        group_element_values = {'group_element': group_element}
-        if group_element.child_leaf:
-            _get_learning_unit_detail(group_element_values, group_element)
-        elif group_element.child_branch:
-            _get_education_group_detail(group_element_values, group_element)
-        group_elements_data.append(group_element_values)
-    return _sorting(group_elements_data)
-
-
-def _sorting(group_elements_data):
-    return sorted(group_elements_data,
-                  key=lambda k: (k.get('group_element').current_order is None,
-                                 k.get('group_element').current_order == -1,
-                                 k.get('group_element').current_order))
-
-
-def _get_education_group_detail(dict_param, group_element):
-    dict_param.update({CODE_SCS: group_element.child_branch.partial_acronym,
-                       TITLE: group_element.child_branch.title,
-                       CREDITS_MIN: group_element.min_credits,
-                       CREDITS_MAX: group_element.max_credits,
-                       BLOCK: None})
-    return dict_param
-
-
-def _get_learning_unit_detail(dict_param, group_element):
-    dict_param.update({CODE_SCS: group_element.child_leaf.acronym,
-                       TITLE: group_element.child_leaf.specific_title,
-                       CREDITS_MIN: None,
-                       CREDITS_MAX: None,
-                       BLOCK: group_element.block,
-                       SESSIONS_DEROGATION: group_element.sessions_derogation})
-    return dict_param
 
 
 def find_root_by_name(text_label_name):
@@ -463,7 +221,7 @@ def education_group_year_admission_condition_edit(request, education_group_year_
     education_group_year = get_object_or_404(EducationGroupYear, pk=education_group_year_id)
 
     education_group_year_root_id = request.GET.get('root')
-    parent = _get_education_group_root(education_group_year_root_id, education_group_year)
+    parent = get_education_group_root(education_group_year_root_id, education_group_year)
 
     acronym = education_group_year.acronym.lower()
